Moved search index from coding to indexer.
