+++ indexer/succinct_trie_builder.hpp
