--- search/search_trie.cpp
@@ -1,5 +1,0 @@
-#include "search/search_trie.hpp"
-#include "search"
-#include "coding/writer.hpp"
-
-void Build(FeaturesVector const & featuresVector, Writer & writer);
